Updated events for nextButton to be triggered from an enhancer.
